[WIP] Add non-working gcc-2.95.3 derivation
GCC-2.95.3 is needed to allow us to both
  * compile the C files generated by the initial GHC build phase
  * compile STGHugs, the GHC-compatible Haskell interpreter included
    with GHC 4
Newer GCC versions reject these old C programs.

This adds a derivation for building GCC-2.95.3 on 32-bit Linux (taken
from Guix, where they did the hard work to make it compile,) which we
can then use in the shell for our build steps for bootstrapping GHC 4.

This derivation has an issue with its linker that means it can't
actually be used to build STGHugs. This needs to be fixed before it will
be really useful.